requirements.txt
webrtcvad==2.0.10
PyAudio==0.2.13
SpeechRecognition==3.8.1
pyttsx3==2.90
Pillow==9.5.0
pycaw
comtypes
